Eliminate compiler warning for gcc x86_64
git-svn-id: file:///home/jj/hercules.svn/trunk@5312 956126f8-22a0-4046-8f4a-272fa8102e63

+++ sllib.c
@@ -947,7 +947,7 @@ sl_vol( SLLABEL *lab,
     }
 
     len = strlen( volser );
-    if( ( len > 6 ) || ( (int)strspn( volser, sl_cset ) != len ) )
+    if( ( len > 6 ) || ( strspn( volser, sl_cset ) != len ) )
     {
         return( SLE_VOLSER );
     }
